+++ skops/hub_utils/__init__.py
@@ -2,10 +2,8 @@ from ._hf_hub import (
     add_files,
     download,
     get_config,
-    get_model_output,
     get_requirements,
     init,
-    push,
     update_env,
 )
 
@@ -14,8 +12,6 @@ __all__ = [
     "download",
     "get_config",
     "get_requirements",
-    "get_model_output",
     "init",
-    "push",
     "update_env",
 ]